Split up the data primitives into respective files
