+++ src/gdt/core/data_primitives/bins.py
@@ -1,6 +1,6 @@
 # CONTAINS TECHNICAL DATA/COMPUTER SOFTWARE DELIVERED TO THE U.S. GOVERNMENT WITH UNLIMITED RIGHTS
 #
-# Contract No.: CA 80MSFC17M0022
+# Contract Nos.: CA 80MSFC17M0022 / 80NSSC24M0035
 # Contractor Name: Universities Space Research Association
 # Contractor Address: 7178 Columbia Gateway Drive, Columbia, MD 21046
 #
@@ -27,927 +27,10 @@
 # License.
 #
 import numpy as np
-from scipy.interpolate import interp1d
-from scipy.integrate import trapezoid
-import copy
+from .intervals import Ebounds
 
-__all__ = ['Range', 'TimeRange', 'EnergyRange', 'Intervals', 'Gti', 'Ebounds',
-           'EventList', 'Bins', 'ExposureBins', 'ChannelBins', 'TimeBins',
-           'EnergyBins', 'TimeChannelBins', 'TimeEnergyBins', 'ResponseMatrix',
-           'Parameter']
-
-class Range():
-    """A primitive class defining a range
-    
-    Parameters:
-        low (float): The low end of the range
-        high (float): The high end of the range
-    """
-    def __init__(self, low, high):
-        if high >= low:
-            self._low = low
-            self._high = high
-        else:
-            self._low = high
-            self._high = low
-
-    @property
-    def center(self):
-        """(float): The center of the range"""
-        return (self._high + self._low) / 2.0
-
-    @property
-    def width(self):
-        """(float): The width of the range"""
-        return self._high - self._low
-
-    def as_tuple(self):
-        """Return the range as a tuple.
-        
-        Returns:  
-            (float, float)
-        """
-        return (self._low, self._high)
-
-    def contains(self, value, inclusive=True):
-        """Determine if the range contains a value.
-        
-        Args:
-            value (float): The input value to check
-            inclusive (bool, optional): 
-                If True, then includes the edges of the range for the check, 
-                otherwise it is edge-exclusive. Default is True.
-        
-        Returns:           
-            bool: True if the value is in the range, False otherwise
-        """
-        if inclusive:
-            test = (value <= self._high) and (value >= self._low)
-        else:
-            test = (value < self._high) and (value > self._low)
-
-        if test:
-            return True
-        else:
-            return False
-
-    @classmethod
-    def intersection(cls, range1, range2):
-        """Return a new Range that is the intersection of two input Ranges.  
-        If the input Ranges do not intersect, then None is returned.
-
-        Args:
-            range1 (:class:`Range`): A range used to calculate the intersection
-            range2 (:class:`Range`): Another range used to calculate the 
-                                     intersection
-        
-        Returns:           
-            :class:`Range`: The intersected range
-        """
-        # test if one low is inside the other range
-        if range1.contains(range2._low):
-            lower = range1
-            upper = range2
-        elif range2.contains(range1._low):
-            lower = range2
-            upper = range1
-        else:
-            return None
-
-        # do the merge
-        low = upper._low
-        high = np.min((lower._high, upper._high))
-        obj = cls(low, high)
-        return obj
-
-    @classmethod
-    def union(cls, range1, range2):
-        """Return a new Range that is the union of two input Ranges
-        
-        Args:
-            range1 (:class:`Range`): A range used to calculate the union
-            range2 (:class:`Range`): Another range used to calculate the union
-        
-        Returns:           
-            :class:`Range`: The unionized range
-        """
-        low = np.min((range1._low, range2._low))
-        high = np.max((range1._high, range2._high))
-        obj = cls(low, high)
-        return obj
-
-    def translate(self, value):
-        """Returns a new Range that is the translated (shifted) by the given value
-
-        Args:
-            value (float): The value to shift the range by
-
-        Returns:
-            :class:`Range`: The shifted range
-        """
-        return self.__class__(self._low + value, self._high + value)
-
-    def __eq__(self, other):
-        return (self._low == other._low) and (self._high == other._high)
-
-    def __repr__(self):
-        return '<{0}: ({1}, {2})>'.format(self.__class__.__name__, self._low,
-                                         self._high)
-
-    def _assert_float(self, value):
-        try:
-            value = float(value)
-        except:
-            raise TypeError('value must be a float')
-        return value
-
-# mark: TODO: add units to TimeRange/EnergyRange
-class TimeRange(Range):
-    """A primitive class defining a time range
-    
-    Parameters:
-        tstart (float): The start time of the range
-        tstop (float): The end time of the range
-    """
-    def __init__(self, tstart, tstop):
-        tstart = self._assert_float(tstart)
-        tstop = self._assert_float(tstop)
-        super().__init__(tstart, tstop)
-
-    @property
-    def duration(self):
-        """(float): The duration of the time range. 
-        Alias for ``width``."""
-        return self.width
-
-    @property
-    def tstart(self):
-        """(float): The start time of the range"""
-        return self._low
-
-    @property
-    def tstop(self):
-        """(float): The end time of the range"""
-        return self._high
-
-
-class EnergyRange(Range):
-    """A primitive class defining an energy range
-    
-    Parameters:
-        emin (float): The low end of the energy range
-        emax (float): The high end of the energy range
-    """
-    def __init__(self, emin, emax):
-        emin = self._assert_float(emin)
-        emax = self._assert_float(emax)
-        super().__init__(emin, emax)
-
-    @property
-    def emax(self):
-        """(float): The high end of the energy range"""
-        return self._high
-
-    @property
-    def emin(self):
-        """(float): The low end of the energy range"""
-        return self._low
-
-    @property
-    def log_center(self):
-        """log_center (float): The logarithmic center of the energy range"""
-        return np.sqrt(self.emin * self.emax)
-
-
-class Intervals():
-    """A primitive class defining a set of intervals or ranges.
-    
-    An interval may be accessed by using indices.
-    
-    Parameters:
-        interval (:class:`Range`, optional): An interval to initialize with
-    """
-    _range_class = Range
-    def __init__(self, interval=None):
-        self._intervals = None
-        if interval is not None:
-            self._intervals = [interval]
-
-    @property
-    def intervals(self):
-        """(list): The list of intervals"""
-        return self._intervals
-
-    @property
-    def num_intervals(self):
-        """(int): The number of intervals"""
-        return len(self._intervals)
-
-    @property
-    def range(self):
-        """(float, float): The full range spanned by the intervals"""
-        if self.num_intervals > 0:
-            return (self._intervals[0].as_tuple()[0],
-                    self._intervals[-1].as_tuple()[1])
-
-    def as_list(self):
-        """Return the intervals as a list of tuples.
-        
-        Returns:     
-            [(float, float), ...]
-        """
-        interval_list = [interval.as_tuple() for interval in self._intervals]
-        return interval_list
-
-    def contains(self, value, inclusive=True):
-        """Determine if the intervals contains a value.
-        
-        Args:
-            value (float): The input value to check
-            inclusive (bool, optional): 
-                If True, then includes the edges of the range for the check, 
-                otherwise it is edge-exclusive. Default is True.
-        
-        Returns:           
-            bool: True if the value is in the intervals, False otherwise
-        """
-        test = [interval.contains(value, inclusive=inclusive) for interval \
-                in self._intervals]
-        return any(test)
-
-    def high_edges(self):
-        """Return a list of the high edges.
-        
-        Returns:     
-            (list)
-        """
-        edges = [interval._high for interval in self._intervals]
-        return edges
-
-    def index(self, value):
-        """Return the index of the interval that contains the value.
-        
-        Note :
-            If the value is precisely on the boundary of two intervals, the
-            first interval index will be returned.
-        
-        Args:
-            value (float): The input value
-        
-        Returns:
-            (int)
-        """
-        for i, interval in enumerate(self._intervals):
-            if interval.contains(value, inclusive=True):
-                return i
-
-    def insert(self, interval):
-        """Insert a new interval
-        
-        Args:
-            interval (:class:`Range`): The interval to insert
-        """
-
-        # if interval is a duplicate, then skip
-        for _interval in self._intervals:
-            if interval == _interval:
-                return
-
-        # where the new range should be inserted
-        idx = [i for i, j in enumerate(self._intervals) if
-               j._low <= interval._low]
-
-        # determine if there is overlap with the lower bounding range, and if 
-        # so, then merge
-        if len(idx) != 0:
-            idx = idx[-1]
-            if self._intervals[idx].contains(interval._low, inclusive=False):
-                the_range = self._intervals.pop(idx)
-                interval = type(interval).union(the_range, interval)
-            else:
-                idx += 1
-        else:
-            idx = 0
-
-        # determine if there is overlap with the upper bounding range, and if 
-        # so, then merge
-        if idx < len(self._intervals):
-            if self._intervals[idx].contains(interval._high):
-                the_range = self._intervals.pop(idx)
-                interval = type(interval).union(the_range, interval)
-
-        self._intervals.insert(idx, interval)
-
-    def low_edges(self):
-        """Return a list of the low edges.
-        
-        Returns:     
-            (list)
-        """
-        edges = [interval._low for interval in self._intervals]
-        return edges
-
-    @classmethod
-    def from_bounds(cls, low_bounds, high_bounds):
-        """Create a new Intervals object from a list of lower and upper bounds.
-        
-        Args:
-            low_bounds (list): The lower bounds of the intervals
-            high_bounds (list): The upper bounds of the intervals
-        
-        Returns:           
-            :class:`Intervals`
-        """
-        low_bounds = np.asarray(low_bounds).flatten()
-        high_bounds = np.asarray(high_bounds).flatten()
-        num = low_bounds.size
-        if num != high_bounds.size:
-            raise ValueError('low_bounds and high_bounds must be of same size')
-
-        intervals = [cls._range_class(low_bounds[i], high_bounds[i]) \
-                     for i in range(num)]
-        obj = cls(intervals[0])
-        obj._intervals = intervals
-        return obj
-
-    @classmethod
-    def from_list(cls, interval_list):
-        """Create a new Intervals object from a list of tuples.
-        
-        Args:
-            interval_list ([(float, float), ...]):  A list of interval tuples
-        
-        Returns:           
-            :class:`Intervals`
-        """
-        intervals = [cls._range_class(*interval) for interval in interval_list]
-        obj = cls(intervals[0])
-        obj._intervals = intervals
-        return obj
-
-    @classmethod
-    def intersection(cls, intervals1, intervals2):
-        """Return a new Intervals object that is the intersection of two 
-        existing Intervals objects.
-        
-        Args:
-            intervals1 (:class:`Intervals`): Intervals to be intersected
-            intervals2 (:class:`Intervals`): Intervals to be intersected
-           
-        Returns:
-            :class:`Intervals`
-        """
-        if intervals1.range[0] <= intervals2.range[0]:
-            _intervals1 = intervals1
-            _intervals2 = intervals2
-        else:
-            _intervals1 = intervals2
-            _intervals2 = intervals1
-
-        new_intervals = []
-        for interval1 in _intervals1.intervals:
-            for interval2 in _intervals2.intervals:
-                if not interval1.contains(interval2._low) and \
-                   not interval1.contains(interval2._high):
-                    continue
-                new_intervals.append(cls._range_class.intersection(interval1,
-                                                                   interval2))
-        obj = cls()
-        obj._intervals = new_intervals
-        return obj
-
-    @classmethod
-    def merge(cls, intervals1, intervals2):
-        """Return a new Intervals object that is a merge of two existing 
-        Intervals objects.
-        
-        Args:
-            intervals1 (:class:`Intervals`): Intervals to be merged
-            intervals2 (:class:`Intervals`): Intervals to be merged
-           
-        Returns:
-            (:class:`Intervals`)
-        """
-        interval_list = intervals1.as_list()
-        interval_list.extend(intervals2.as_list())
-        interval_list = sorted(interval_list)
-        obj = cls(cls._range_class(*interval_list.pop(0)))
-        for interval in interval_list:
-            obj.insert(cls._range_class(*interval))
-        return obj
-
-    def __repr__(self):
-        s = '<{0}: {1} intervals; range {2}>'.format(self.__class__.__name__,
-                                                     self.num_intervals,
-                                                     self.range)
-        return s
-
-    def __getitem__(self, index):
-        if index > self.num_intervals-1:
-            raise KeyError('Requested {0}th interval and only {1} intervals '\
-                           'exist'.format(index, self.num_intervals))
-        return self._intervals[index]
-
-
-class Gti(Intervals):
-    """A primitive class defining a set of Good Time Intervals (GTIs).
-    
-    An interval may be accessed by using indices.
-    
-    Parameters:
-        interval (:class:`TimeRange`, optional): An interval to initialize with
-    """
-    _range_class = TimeRange
-
-    @classmethod
-    def from_boolean_mask(cls, times, mask):
-        """Create a new GTI object from a list of times and a Boolean mask
-        Splits the boolean mask into segments of contiguous values and applies
-        to array of times to create a GTI object.
-        
-        Args:
-            times (np.array): An array of times
-            mask (np.array(dtype=bool)): The boolean array. Must be the same 
-                                         size as times.
-        
-        Returns:           
-            :class:`Gti`: The new GTI object
-        """
-        times = np.asarray(times)
-        mask = np.asarray(mask)
-
-        # split a boolean mask array into segments based on True/False
-        indices = np.nonzero(mask[1:] != mask[:-1])[0] + 1
-        time_segs = np.split(times, indices)
-        mask_segs = np.split(mask, indices)
-
-        # retrieve the start and stop times for the "on" intervals
-        segs = []
-        numsegs = len(indices) + 1
-        for i in range(numsegs):
-            if mask_segs[i][0]:
-                segs.append((time_segs[i][0], time_segs[i][-1]))
-
-        # if mask is all True or all False
-        if len(segs) == 0:
-            if mask[0]:
-                segs = [(times.min(), times.max())]
-            else:
-                return None
-
-        return cls.from_list(segs)
-
-
-class Ebounds(Intervals):
-    """A primitive class defining a set of energy bounds.
-    
-    An interval may be accessed by using indices.
-    
-    Parameters:
-        interval (:class:`EnergyRange`, optional): An interval to initialize with
-    """
-    _range_class = EnergyRange
-
-
-class EventList():
-    """A primitive class defining an event list.
-    
-    Parameters:
-        times (np.array): The array of event times
-        channels (np.array): The corresponding array of associated energy channel
-        ebounds (:class:`Ebounds`, optional): The energy bounds mapped to 
-                                              energy channels
-    """
-    def __init__(self, times=None, channels=None, ebounds=None):
-
-        if (times is not None) and (channels is not None):
-            times = np.asarray(times).flatten()
-            channels = np.asarray(channels).flatten()
-            if times.size != channels.size:
-                raise ValueError('times and channels arrays must be same length')
-        else:
-            times = np.array([], dtype=float)
-            channels = np.array([], dtype=int)
-
-        self._events = np.empty(times.size, dtype=[('TIME', times.dtype.type),
-                                                   ('PHA', channels.dtype.type)])
-        self._events['TIME'] = times
-        self._events['PHA'] = channels
-
-        if ebounds is not None:
-            if not isinstance(ebounds, Ebounds):
-                raise TypeError('ebounds must be of type Ebounds')
-        self._ebounds = ebounds
-
-    @property
-    def channel_range(self):
-        """(int, int): The range of the channels in the list"""
-        if self.size > 0:
-            return int(np.min(self.channels)), int(np.max(self.channels))
-
-    @property
-    def channels(self):
-        """(np.array): The PHA channel array"""
-        return self._events['PHA']
-
-    @property
-    def ebounds(self):
-        """(:class:`Ebounds`): The energy bounds of the energy channels.
-                               This property can be set.
-        """
-        return self._ebounds
-
-    @ebounds.setter
-    def ebounds(self, val):
-        if not isinstance(val, Ebounds):
-            raise TypeError('ebounds must be an Ebounds object')
-        self._ebounds = val
-
-    @property
-    def emax(self):
-        """(float): The maximum energy"""
-        if self._ebounds is not None:
-            return self._ebounds.range[1]
-
-    @property
-    def emin(self):
-        """(float): The minimum energy"""
-        if self._ebounds is not None:
-            return self._ebounds.range[0]
-
-    @property
-    def energy_range(self):
-        """(float, float): The energy range of the channels"""
-        if self.size > 0:
-            if self._ebounds is not None:
-                chan_min, chan_max = self.channel_range
-                return (self._ebounds[chan_min].emin,
-                        self._ebounds[chan_max].emax)
-
-    @property
-    def num_chans(self):
-        """(int): The number of energy channels. 
-        Note that not all channels will necessarily have events, especially if a 
-        slice is made over energy."""
-        if self.ebounds is not None:
-            return self._ebounds.num_intervals
-        else:
-            return (self._events['PHA'].max() - self._events['PHA'].min()) + 1
-
-    @property
-    def size(self):
-        """(int): The number of events in the list"""
-        return self._events.size
-
-    @property
-    def time_range(self):
-        """(float, float): The range of the times in the list"""
-        if self.size > 0:
-            return self.times.min(), self.times.max()
-
-    @property
-    def times(self):
-        """(np.array): The event times"""
-        return self._events['TIME']
-
-    def bin(self, method, *args, tstart=None, tstop=None,
-            event_deadtime=2.6e-6, overflow_deadtime=1.0e-5, **kwargs):
-        """Bin the EventList in time given a binning function and return a
-        2D time-energy channel histogram. If the ebounds energy calibration is
-        set, returns a :class:`TimeEnergyBins` object, otherwise returns a
-        :class:`TimeChannelBins` object.
-        
-        The binning function should take as input an array of times as well
-        as a tstart and tstop keywords for partial list binning.  Additional 
-        arguments and keyword arguments specific to the function are allowed.
-        The function should return an array of time edges for the bins, such
-        that, for `n` bins, there are `n` + 1 edges.
-        
-        Args:
-            method (<function>): A binning function
-            *args: Arguments to be passed to the binning function
-            tstart (float, optional): 
-                If set, defines the start time of the EventList to be binned, 
-                otherwise binning will begin at the time of the first event.
-            tstop (float, optional): 
-                If set, defines the end time of the EventList to be binned, 
-                otherwise binning will end at the time of the last event.
-            event_deadtime (float, optional): The deadtime per event in seconds. 
-                                              Default is 2.6e-6.
-            overflow_deadtime (float, optional): 
-                The deadtime per event in the overflow channel in seconds. 
-                Default is 1e-5.
-            **kwargs: Options to be passed to the binning function
-        
-        Returns:
-            (:class:`TimeEnergyBins` or :class:`TimeChannelBins`)
-        """
-        if tstart is None:
-            tstart = self.time_range[0]
-        if tstop is None:
-            tstop = self.time_range[1]
-
-        # set the start and stop of the rebinning segment        
-        mask = (self.times >= tstart) & (self.times <= tstop)
-        bin_times = self.times[mask]
-        kwargs['tstart'] = tstart
-        kwargs['tstop'] = tstop
-
-        # get the time edges from the binning function and then do the 2d histo
-        time_edges = method(bin_times, *args, **kwargs)
-        if self.ebounds is not None:
-            chan_list = np.arange(self.ebounds.num_intervals + 1)
-        else:
-            chan_list = np.arange(self.channel_range[1] + 2)
-        counts = np.histogram2d(bin_times, self.channels[mask],
-                                [time_edges, chan_list])[0]
-
-        # calculate exposure
-        lo_edges = time_edges[:-1]
-        hi_edges = time_edges[1:]
-        overflow_counts = counts[:, -1]
-        deadtime = counts.sum(axis=1) * event_deadtime + \
-                   overflow_counts * (overflow_deadtime - event_deadtime)
-        exposure = (hi_edges - lo_edges) - deadtime
-
-        if self.ebounds is None:
-            bins = TimeChannelBins(counts, lo_edges, hi_edges, exposure,
-                                   chan_list[:-1])
-        else:
-            bins = TimeEnergyBins(counts, lo_edges, hi_edges, exposure,
-                                  self.ebounds.low_edges(),
-                                  self.ebounds.high_edges())
-        return bins
-
-    def channel_slice(self, chanlo, chanhi):
-        """Perform a slice in energy channels of the EventList and return a 
-        new EventList
-        
-        Args:
-            chanlo (int): The start of the channel slice
-            chanhi (int): The end of the channel slice
-        
-        Returns:        
-            (:class:`EventList`)
-        """
-        mask = (self.channels >= chanlo) & (self.channels <= chanhi)
-        events = self._events[mask]
-        return EventList(times=events['TIME'], channels=events['PHA'],
-                         ebounds=copy.deepcopy(self._ebounds))
-
-    def count_spectrum(self, **kwargs):
-        """Extract a count spectrum for the EventList or for a segment of the
-        EventList. If the ebounds energy calibration is set, returns an 
-        :class:`EnergyBins` object, otherwise returns a :class:`ChannelBins`
-        object.
-        
-        Args:
-            time_ranges ([(float, float), ...], optional): 
-                The time range or time ranges over which to calculate the 
-                exposure. If omitted, calculates the total exposure of the data.
-            event_deadtime (float, optional): The deadtime per event in seconds. 
-                                              Default is 0.
-            overflow_deadtime (float, optional): 
-                The deadtime per event in the overflow channel in seconds. 
-                Default is 0.
-        
-        Returns:        
-            (:class:`EnergyBins` or :class:`ChannelBins`)
-        """
-        if self.ebounds is not None:
-            chan_list = np.arange(self.ebounds.num_intervals + 1)
-        else:
-            chan_list = np.arange(self.channel_range[1] + 2)
-        counts = np.histogram(self.channels, bins=chan_list)[0]
-        exposure = np.full(chan_list.size-1, self.get_exposure(time_ranges=None,
-                                                               **kwargs))
-
-        if self.ebounds is None:
-            bins = ChannelBins.create(counts, chan_list[:-1], exposure)
-        else:
-            bins = EnergyBins(counts, self.ebounds.low_edges(),
-                              self.ebounds.high_edges(), exposure)
-        return bins
-
-    def energy_slice(self, emin, emax):
-        """Perform a slice in energy of the EventList and return a new EventList.
-        Since energies are binned, an emin or emax falling inside of an energy
-        channel bin will include that bin in the slice.
-        
-        Args:
-            emin (float): The start of the energy slice
-            emax (float): The end of the energy slice
-        
-        Returns:        
-            (:class:`EventList`)
-        """
-        if self._ebounds is None:
-            raise AttributeError('Ebounds has not been set, therefore the \
-                                  the EventList cannot be sliced by energy')
-        emins = np.array([ebound.emin for ebound in self._ebounds.intervals])
-        emaxs = np.array([ebound.emax for ebound in self._ebounds.intervals])
-        mask = (emins < emax) & (emaxs > emin)
-        chan_nums = np.arange(self.channel_range[0], self.channel_range[1]+1)
-        chan_nums = chan_nums[mask]
-        return self.channel_slice(chan_nums[0], chan_nums[-1])
-
-    def get_exposure(self, time_ranges=None, event_deadtime=0.0,
-                     overflow_deadtime=0.0):
-        """Calculate the total exposure, in seconds, of a time range or time 
-        ranges of data.
-        
-        Args:
-            time_ranges ([(float, float), ...], optional): 
-                The time range or time ranges over which to calculate the 
-                exposure. If omitted, calculates the total exposure of the data.
-            event_deadtime (float, optional): The deadtime per event in seconds. 
-                                              Default is 0.
-            overflow_deadtime (float, optional): 
-                The deadtime per event in the overflow channel in seconds. 
-                Default is 0.
-        
-        Returns:
-            (float)
-        """
-        if time_ranges is None:
-            time_ranges = [self.time_range]
-        try:
-            iter(time_ranges[0])
-        except:
-            time_ranges = [time_ranges]
-
-        exposure = 0.0
-        for i in range(len(time_ranges)):
-            tstart, tstop = self._assert_range(time_ranges[i])
-            tcent = (tstop + tstart) / 2.0
-            dt = (tstop - tstart) / 2.0
-            mask = (np.abs(self.times - tcent) <= dt)
-            # mask = (self.time >= tstart) & (self.time < tstop)
-            deadtime = mask.sum() * event_deadtime
-
-            if self.ebounds is not None:
-                of_chan = self.ebounds.num_intervals-1
-            else:
-                of_chan = self.channel_range[1]
-
-            deadtime += np.sum(self.channels[mask] == of_chan) * \
-                        (overflow_deadtime - event_deadtime)
-            exposure += (tstop - tstart) - deadtime
-        return exposure
-
-    def rebin_energy(self, method, *args, **kwargs):
-        """Rebin the energy channels using the specified binning algorithm.  
-        This does not change the number of events in the EventList, but changes 
-        their assignment to a channel and bins the energy bounds mapping to 
-        those channels (if applicable).  A new EventList object is returned.
-        
-        Args:
-            method (<function>): A binning function
-            *args: Arguments to be passed to the binning function
-            **kwargs: Deadtime arguments passed to get_exposure
-        
-        Returns:
-           (:class:`EventList`)
-        """
-        # exposure and counts; not really used other than for some specific
-        # binning algorithms
-        exposure = np.full(self.num_chans, self.get_exposure(**kwargs))
-        chans, counts = np.unique(self.channels, return_counts=True)
-        if chans.size != self.num_chans:
-            counts_fill = np.zeros(self.num_chans)
-            counts_fill[chans] = counts
-            counts = counts_fill
-        edges = np.arange(self.num_chans + 1)
-
-        # call the binning algorithm and get the new edges
-        _, _, _, new_edges = method(counts, np.sqrt(counts), exposure, edges, *args)
-
-        # re-assign the pha channels based on the new edges
-        # and also rebin the ebounds
-        new_channels = np.zeros_like(self.channels)
-        new_ebounds = []
-        for i in range(len(new_edges) - 1):
-            emin = new_edges[i]
-            emax = new_edges[i + 1]
-            mask = (self.channels >= emin) & (self.channels < emax)
-            new_channels[mask] = i
-            if self.ebounds is not None:
-                new_ebounds.append((self.ebounds[emin].emin,
-                                    self.ebounds[emax - 1].emax))
-
-        # create the new EventList object with the rebinned channels
-        if self.ebounds is not None:
-            new_ebounds = Ebounds.from_list(new_ebounds)
-        else:
-            new_ebounds = None
-        obj = EventList(times=np.copy(self.times), channels=new_channels,
-                        ebounds=new_ebounds)
-        return obj
-
-    def sort_time(self):
-        """In-place sort by time.
-        """
-        idx = np.argsort(self._events['TIME'])
-        self._events = self._events[:][idx]
-
-    def sort_channels(self):
-        """In-place sort by channel number.
-        """
-        idx = np.argsort(self._events['PHA'])
-        self._events = self._events[:][idx]
-
-    def time_slice(self, tstart, tstop):
-        """Perform a slice in time of the EventList and return a new EventList
-        
-        Args:
-            tstart (float): The start of the time slice
-            tstop (float): The end of the time slice
-        
-        Returns:        
-            (:class:`EventList`)
-        """
-        mask = (self.times >= tstart) & (self.times <= tstop)
-        events = self._events[mask]
-        return EventList(times=events['TIME'], channels=events['PHA'],
-                         ebounds=copy.deepcopy(self._ebounds))
-
-    @classmethod
-    def merge(cls, eventlists, sort=False, force_unique=True):
-        """Merge multiple EventLists together in time and optionally sort.
-        
-        Args:
-            eventlist (list of :class:`EventList`): 
-                A list containing the EventLists to be merged
-            sort (bool, optional): 
-                If True, sorts by time after the merge. Default is False.
-            force_unique (bool, optional): 
-                If True, force all events to be unique via brute force sorting. 
-                If False, the EventLists will only be checked and masked for 
-                overlapping time ranges. Events can potentially be lost if the 
-                merged EventLists contain overlapping times (but not necessarily 
-                duplicate events), however this method is much faster.  
-                Default is True.
-        
-        Returns:
-            (:class:`EventList`)
-        """
-
-        # put in time order
-        idx = np.argsort([eventlist.times.min() for eventlist in eventlists])
-        eventlists = [eventlists[i] for i in idx]
-
-
-        # mark: TODO add check for ebounds consistency
-        ebounds_idx = 0
-        for i in range(len(eventlists)):
-            if eventlists[i].ebounds is None:
-                continue
-            else:
-                ebounds_idx = i
-                break
-
-        new_times = np.array([])
-        new_chans = np.array([])
-        for eventlist in eventlists:
-            # skip empty EventLists
-            if eventlist.size == 0:
-                continue
-
-            # if not forcing to be unique, just make sure there is no time overlap
-            if (not force_unique) and (new_times.size > 0):
-                mask = (eventlist.times > new_times.max())
-                temp_times = eventlist.times[mask]
-                temp_chans = eventlist.channels[mask]
-            else:
-                temp_times = eventlist.times
-                temp_chans = eventlist.channels
-            new_times = np.concatenate((new_times, temp_times))
-            new_chans = np.concatenate((new_chans, temp_chans))
-
-        # force unique: make sure that we only keep unique events (slower)
-        if force_unique:
-            new_times, uidx = np.unique(new_times, return_index=True)
-            new_chans = new_chans[uidx]
-
-        obj = cls(times=new_times, channels=new_chans,
-                  ebounds=copy.deepcopy(eventlists[ebounds_idx].ebounds))
-
-        # do a sort
-        if sort and not force_unique:
-            obj.sort_time()
-
-        return obj
-
-    def _assert_range(self, valrange):
-        assert valrange[0] <= valrange[1], \
-            'Range must be in increasing order: (lo, hi)'
-        return valrange
-
-    def __repr__(self):
-        s = '<EventList: {0} events;\n'.format(self.size)
-        s += ' time range {0};\n channel range {1}>'.format(self.time_range,
-                                                            self.channel_range)
-        return s
+__all__ = ['Bins', 'ExposureBins', 'ChannelBins', 'TimeBins',
+           'EnergyBins', 'TimeChannelBins', 'TimeEnergyBins']
 
 
 class Bins():
@@ -3344,473 +2427,3 @@ class TimeEnergyBins():
             s += ';\n {0} energy segments'.format(len(self._good_energy_segments))
 
         return s+'>'
-
-
-class ResponseMatrix():
-    """A class defining a 2D detector response matrix, with an (input) photon
-    axis and (output) channel axis.
-
-    Parameters:
-        matrix (np.array): The 2D matrix, of shape 
-                           (num_photon_bins, num_channels)
-        emin (np.array): The low edges of the (input) photon bins
-        emax (np.array): The high edges of the (input) photon bins
-        chanlo (np.array): The low edges of the (output) energy channels
-        chanhi (np.array): The high edges of the (output) energy channels    
-    """
-    def __init__(self, matrix, emin, emax, chanlo, chanhi):
-        try:
-            iter(matrix)
-            self._matrix = np.asarray(matrix)
-        except:
-            raise TypeError('matrix must be an iterable')
-        if self._matrix.ndim != 2:
-            raise TypeError('matrix must be a 2-dimensional array')
-
-        try:
-            iter(emin)
-            self._emin = np.asarray(emin).flatten()
-        except:
-            raise TypeError('emin must be an iterable')
-        try:
-            iter(emax)
-            self._emax = np.asarray(emax).flatten()
-        except:
-            raise TypeError('emax must be an iterable')
-
-        try:
-            iter(chanlo)
-            self._chanlo = np.asarray(chanlo).flatten()
-        except:
-            raise TypeError('chanlo must be an iterable')
-        try:
-            iter(chanhi)
-            self._chanhi = np.asarray(chanhi).flatten()
-        except:
-            raise TypeError('chanhi must be an iterable')
-
-        if (self._emin.size != self._emax.size):
-            raise ValueError('emin and emax must be the same length')
-
-        if (self._chanlo.size != self._chanhi.size):
-            raise ValueError('chanlo and chanhi must be the same length')
-
-        if (self._matrix.shape[0] != self._emin.size) or \
-           (self._matrix.shape[1] != self._chanlo.size):
-            raise ValueError('matrix axis 0 must have same length as emin and '\
-                             'emax.  matrix axis 1 must have same length as '\
-                             'chanlo and chanhi.')
-
-    @property
-    def channel_centroids(self):
-        """(np.array): The geometric mean of the energy channels"""
-        return np.sqrt(self._chanlo * self._chanhi)
-
-    @property
-    def channel_widths(self):
-        """(np.array): The energy channel widths"""
-        return self._chanhi - self._chanlo
-
-    @property
-    def ebounds(self):
-        """(:class:`Ebounds`): The energy bounds"""
-        return Ebounds.from_bounds(self._chanlo, self._chanhi)
-
-    @property
-    def matrix(self):
-        """(np.array): The raw matrix"""
-        return self._matrix
-
-    @property
-    def num_chans(self):
-        """(int): The number of energy channels"""
-        return self._matrix.shape[1]
-
-    @property
-    def num_ebins(self):
-        """(int): The number of photon bins"""
-        return self._matrix.shape[0]
-
-    @property
-    def photon_bins(self):
-        """(:class:`Ebounds`): The photon bins"""
-        return Ebounds.from_bounds(self._emin, self._emax)
-
-    @property
-    def photon_bin_centroids(self):
-        """(np.array): The geometric mean of the photon bins"""
-        return np.sqrt(self._emin * self._emax)
-
-    @property
-    def photon_bin_widths(self):
-        """(np.array): The photon bin widths"""
-        return self._emax - self._emin
-
-    def channel_effective_area(self):
-        """Returns the effective area as a function of recorded channel energy
-        (integrated over incident photon bins).
-        
-        Returns:
-            (:class:`Bins`)
-        """
-        bins = Bins(self._matrix.sum(axis=0), self._chanlo, self._chanhi)
-        return bins
-
-    def effective_area(self, energy, interp_kind='linear'):
-        """Calculate the effective area at a given energy or an array of
-        energies.  This function interpolates the DRM (in log10(cm^2)) to
-        provide the effective area for any energy within the photon energy
-        bounds of the DRM.
-        
-        Args:
-            energy (float or np.array): The photon energy or energies at which
-                                        to calculate the effective area.
-            interp_kind (str, optional): The kind of interpolation to be 
-                                         passed to scipy.interp1d.  Default is
-                                         'linear'.
-        
-        Returns:
-            (np.array)
-        """
-        try:
-            energy = np.asarray(energy, dtype=float)
-        except:
-            raise TypeError('energy must be a positive float')
-
-        if np.any(energy < self.photon_bins.range[0]) or \
-           np.any(energy > self.photon_bins.range[1]):
-            raise ValueError('energy must be within the photon energy bounds ' \
-                             'of the DRM')
-
-        effarea = self.photon_effective_area()
-        diff_effarea = effarea.counts
-
-        # create the DRM interpolator
-        x = np.append(self._emin, self._emax[-1])
-        y = np.append(diff_effarea, diff_effarea[-1])
-        nz_mask = (y > 0)
-        effarea_interpolator = interp1d(x[nz_mask], np.log10(y[nz_mask]),
-                                        kind=interp_kind,
-                                        fill_value='extrapolate')
-
-        # do the interpolation
-        effarea = 10.0**effarea_interpolator(energy)
-        return effarea
-
-    def fold_spectrum(self, function, params, channel_mask=None):
-        """Fold a photon spectrum through a DRM to get a count spectrum
-        
-        Args: 
-            function (<function>): 
-                A photon spectrum function.  The function must accept a list of 
-                function parameters as its first argument and an array of photon 
-                energies as its second argument.  The function must return the 
-                evaluation of the photon model in units of ph/s-cm^2-keV.
-            params (list of float): A list of parameter values to be passed to
-                                   the photon spectrum function
-            channel_mask (np.array, optional): 
-                A boolean mask where True indicates the channel is to be used 
-                for folding and False indicates the channel is to not be used 
-                for folding.  If omitted, all channels are used.
-        
-        Returns:        
-            (np.array)
-        """
-        drm = self._matrix
-        if channel_mask is not None:
-            drm = drm[:, channel_mask]
-
-        # evaluate photon model
-        photon_model = function(params, self.photon_bin_centroids)
-
-        # fold photon model through DRM
-        counts = np.dot(drm.T, photon_model * self.photon_bin_widths)
-
-        return counts
-
-    def photon_effective_area(self):
-        """Returns the effective area as a function of incident photon energy
-        (integrated over recorded energy channels).
-        
-        Returns:        
-            (:class:`Bins`)
-        """
-        bins = Bins(self._matrix.sum(axis=1), self._emin, self._emax)
-        return bins
-
-    def rebin(self, factor=None, edge_indices=None):
-        """Rebins the channel energy axis of a DRM and returns a new response
-        object.  Rebinning can only be used to downgrade the channel resolution
-        and is constrained to the channel edges of the current DRM. 
-        
-        Rebinning can be performed by either defining an integral factor of the
-        number of current energy channels to combine (e.g. factor=4 for 128
-        energy channels would result in 32 energy channels), or by defining an
-        index array into the current channel edges that will define the new set
-        of edges.
-        
-        Args:
-            factor (int, optional): The rebinning factor. Must set either this
-                                    or `edge_indices`
-            edge_indices (np.array, optional): The index array that represents
-                                               which energy edges should remain
-                                               in the rebinned DRM.
-        
-        Returns:
-            (:class:`ResponseMatrix`)
-        """
-        if (factor is None) and (edge_indices is None):
-            raise ValueError('Either factor or edge_indices must be set')
-
-        elif factor is not None:
-            try:
-                factor = int(factor)
-            except:
-                raise TypeError('factor must be a positive integer')
-            if factor < 1:
-                raise ValueError('factor must be a positive integer')
-            if (self.num_chans % factor) > 0:
-                raise ValueError('factor must be factor of numchans: '\
-                                 '{}'.format(self.num_chans))
-            chanlo = self._chanlo.reshape(-1, factor)[:,0]
-            chanhi = self._chanhi.reshape(-1, factor)[:,-1]
-
-        elif edge_indices is not None:
-            try:
-                edge_indices = np.asarray(edge_indices)
-            except:
-                raise TypeError('new_edges must be a list, tuple or array')
-            edge_indices = np.unique(edge_indices.astype(int))
-            if (edge_indices[0] < 0) or (edge_indices[-1] > self.num_chans):
-                raise ValueError('edge_indices outside valid range of ' \
-                                 '0-{}'.format(self.num_chans))
-
-            old_edges = np.append(self._chanlo, self._chanhi[-1])
-            new_edges = old_edges[edge_indices]
-            chanlo = new_edges[:-1]
-            chanhi = new_edges[1:]
-
-        # the new effective area matrix will just be summed over the channel
-        # axis for the bins we are combining
-        new_matrix = np.zeros((self.num_ebins, chanlo.size))
-        sidx = (self._chanlo[:,np.newaxis] == chanlo[np.newaxis,:]).nonzero()[0]
-        eidx = (self._chanhi[:,np.newaxis] == chanhi[np.newaxis,:]).nonzero()[0]
-        for i in range(chanlo.size):
-            new_matrix[:,i] = self._matrix[:,sidx[i]:eidx[i]+1].sum(axis=1)
-
-        obj = type(self)(new_matrix, self._emin, self._emax, chanlo, chanhi)
-        return obj
-
-    def resample(self, num_photon_bins=None, photon_bin_edges=None,
-                 num_interp_points=20, interp_kind='linear'):
-        """Resamples the incident photon axis of a DRM and returns a new 
-        ResponseMatrix object.  Resampling can be used to downgrade the photon 
-        energy resolution, upgrade the resolution, and/or redefine the edges of 
-        the incident photon bins.  By definition, the resampling can only be 
-        performed within the photon energy range of the current object.
-        
-        The process for resampling is to create a high-resolution grid for each
-        new photon bin, interpolate the differential effective area onto that
-        grid (interpolation is performed on log10(effective area)), and then
-        integrate the differential effective area over the grid points in each
-        bin.  The final effective area is then scaled by the ratio of the 
-        initial number of photon bins to the requested number of photon bins.
-        
-        Args:
-            num_photon_bins (int, optional): The number of photon bins in the
-                                             new DRM. The bin edges will be 
-                                             generated logarithmically. Only set
-                                             this or `photon_bin_edges`.
-            photon_bin_edges (np.array, optional): The array of photon bin edges.
-                                                   Only set this or 
-                                                   `num_photon_bins`
-            num_interp_points (int, optional): The number of interpolation points
-                                               used to integrate over for each
-                                               new photon bin. Default is 20.
-            interp_kind (str, optional): The kind of interpolation to be 
-                                         passed to scipy.interp1d.  Default is
-                                         'linear'.
-        
-        Returns:
-            (:class:`ResponseMatrix`)
-        """
-        if (num_photon_bins is None) and (photon_bin_edges is None):
-            raise ValueError('Either num_photon_bins or photon_bin_edges must' \
-                             ' be set')
-        elif num_photon_bins is not None:
-            try:
-                num_photon_bins = int(num_photon_bins)
-                assert num_photon_bins > 0
-            except:
-                raise TypeError('num_photon_bins must be a positive integer')
-            photon_bin_edges = np.geomspace(self._emin[0], self._emax[-1],
-                                            num_photon_bins+1)
-
-        elif photon_bin_edges is not None:
-            try:
-                photon_bin_edges = np.asarray(photon_bin_edges)
-            except:
-                raise TypeError('photon_bin_edges must be an iterable')
-            badmask = (photon_bin_edges < self._emin[0]) | \
-                      (photon_bin_edges > self._emax[-1])
-            if badmask.sum() > 0:
-                raise ValueError('photon_bin_edges are beyond valid range of '\
-                                 '{0:.2f}-{1:.2f}'.format(self._emin[0],
-                                                          self._emax[-1]))
-            photon_bin_edges = np.sort(photon_bin_edges)
-            num_photon_bins = photon_bin_edges.size-1
-
-        # differential effective area
-        diff_effarea = self._matrix/self.photon_bin_widths[:,np.newaxis]
-
-        # create an interpolation array over each new bin
-        photon_bounds = np.vstack((photon_bin_edges[:-1], photon_bin_edges[1:]))
-        interp_arr = np.geomspace(*photon_bounds, num_interp_points+2, axis=0)
-
-        # create the DRM interpolator
-        x = np.append(self._emin, self._emax[-1])
-        y = np.vstack((diff_effarea, diff_effarea[-1,:]))
-        y[y == 0] = 1e-10
-        effarea_interpolator = interp1d(x, np.log10(y), axis=0,
-                                        kind=interp_kind,
-                                        fill_value='extrapolate')
-
-        # interpolate the DRM over the photon interpolation array
-        diff_effarea_interp = 10.**effarea_interpolator(interp_arr)
-        badmask = np.isnan(diff_effarea_interp)
-        diff_effarea_interp[badmask] = 0.0
-
-        # integrate the DRM over the photon interpolation array
-        diff_effarea_new = trapezoid(diff_effarea_interp,
-                                 interp_arr[:,:,np.newaxis], axis=0)
-
-        # scale by ratio of photon bins
-        drm = diff_effarea_new / (self.num_ebins/num_photon_bins)
-
-        # create response object
-        obj = type(self)(drm, photon_bin_edges[:-1], photon_bin_edges[1:],
-                         self._chanlo, self._chanhi)
-        return obj
-
-    def __repr__(self):
-        return '<ResponseMatrix: {0} energy bins; {1} ' \
-               'channels>'.format(self.num_ebins, self.num_chans)
-
-
-class Parameter():
-    """A fit parameter class
-    
-    Parameters:
-        value (float): The central fit value
-        uncert (float or 2-tuple): The 1-sigma uncertainty. If a 2-tuple, then 
-                                   is of the form (low, high)
-        name (str, optional): The name of the parameter
-        units (str, optional): The units of the parameter
-        support (2-tuple, optional): The valid support of the parameter
-    """
-    def __init__(self, value, uncert, name='', units=None,
-                 support=(-np.inf, np.inf)):
-
-        self._value = float(value)
-        if isinstance(uncert, (tuple, list)):
-            if len(uncert) == 2:
-                pass
-            elif len(uncert) == 1:
-                uncert = (uncert[0], uncert[0])
-            else:
-                raise ValueError('uncertainty must be a 1- or 2-tuple')
-        elif isinstance(uncert, float):
-            uncert = (uncert, uncert)
-        else:
-            raise TypeError('uncertainty must be a float or 1- or 2-tuple')
-        self._uncert = uncert
-
-        self._units = units
-        self._name = name
-        self._support = support
-
-    @property
-    def name(self):
-        """(str): The name of the parameter"""
-        return self._name
-
-    @property
-    def support(self):
-        """(2-tuple): The valid support of the parameter"""
-        return self._support
-
-    @property
-    def uncertainty(self):
-        """(float, float): The 1-sigma uncertainty"""
-        return self._uncert
-
-    @property
-    def units(self):
-        """(str): The units of the parameter"""
-        return self._units
-
-    @property
-    def value(self):
-        """(float): The central fit value"""
-        return self._value
-
-    def one_sigma_range(self):
-        """Return the 1 sigma range of the parameter fit
-        
-        Returns:
-            (tuple): 2-tuple (low, high)
-        """
-        return (self.value - self.uncertainty[0],
-                self.value + self.uncertainty[1])
-
-    def to_fits_value(self):
-        """Return as a tuple to be used for a FITS file
-
-        Returns:
-            (tuple): 3-value tuple (value, +uncertainty, -uncertainty)   
-        """
-        return (self.value, *self.uncertainty[::-1])
-
-    def valid_value(self):
-        """Check if the parameter value is within the allowed parameter range
-        
-        Returns:
-            (bool)
-        """
-        if (self.value >= self.support[0]) and \
-                (self.value <= self.support[1]):
-            return True
-        else:
-            return False
-
-    def _str_format(self):
-        if (self.value > 0.005) and (self.uncertainty[0] > 0.005):
-            value = '{0:.2f}'.format(self.value)
-            uncertainty = tuple(
-                ['{0:.2f}'.format(u) for u in self.uncertainty])
-        else:
-            value = '{0:.2e}'.format(self.value)
-            val_coeff, val_exp = value.split('e')
-            val_exp = int(val_exp)
-            uncertainty = ['{0:.2e}'.format(u) for u in self.uncertainty]
-            uncert_coeff = []
-            uncert_exp = []
-            for uncert in uncertainty:
-                uncert_coeff.append(uncert.split('e')[0])
-                uncert_exp.append(int(uncert.split('e')[1]))
-        return (value, uncertainty)
-
-    def __repr__(self):
-        return '<{0}: {1}>'.format(self.__class__.__name__, self.name)
-
-    def __str__(self):
-        value, uncertainty = self._str_format()
-
-        if uncertainty[0] == uncertainty[1]:
-            s = '+/- {0}'.format(uncertainty[0])
-        else:
-            s = '+{1}/-{0}'.format(uncertainty[0], uncertainty[1])
-        if self.units is None:
-            return '{0}: {1} {2}'.format(self.name, value, s)
-        else:
-            return '{0}: {1} {2} {3}'.format(self.name, value, s, self.units)
